// src/models/ComponentAliasEnums.ts
export enum COMPONENT_ALIAS_ID {
  /** Describes the employee trend */
  RDemEmpTrend = "RDemEmpTrend",
  RCompSplit = "RCompSplit",
  EmployeeHireSummaryCard = "EmployeeHireSummaryCard",
  RDeductionsTrend = "RDeductionsTrend",
  RDemGender = "RDemGender",
  RDemBusinessUnit = "RDemBusinessUnit",
  RDemContractType = "RDemContractType",
  RCompNonPayrollSpending = "RCompNonPayrollSpending",
  TimePerGradePromotion = "TimePerGradePromotion",
  PromotionsSummaryCard = "PromotionsSummaryCard",
  SanctionsSummaryCard = "SanctionsSummaryCard",
  TimeToHireStrata = "TimeToHireStrata",
}